player plays. state bug

+++ chess/csharp_chess/cschess.agents.test/UniformEvaluator.cs
@@ -4,11 +4,11 @@ namespace cschess.agents.test;
 
 internal class UniformEvaluator : IEvaluator
 {
-    public IEnumerable<(Dictionary<Move, double>, double)> BatchEval(IEnumerable<IChessGame> games)
+    public IEnumerable<(Dictionary<Move, float>, float)> BatchEval(IEnumerable<IChessGame> games)
     {
         return from chessGame in games
             select chessGame.LegalMoves().ToList() into moves
-            let prob = 1.0 / moves.Count
-            select (moves.ToDictionary(x => x, _ => prob), 0.0);
+            let prob = 1.0f / moves.Count
+            select (moves.ToDictionary(x => x, _ => prob), 0.0f);
     }
 }
